release 1.0.2

+++ checks/data_plausibility_checks/utils/data.py
@@ -81,17 +81,17 @@ def check_variable_conditions(
                 else:
                     coords_tuple = tuple(int(c) for c in indices + tuple(coord))
 
-                value = int(values[tuple(coord)] if values is not None else None)
+                value = float(values[tuple(coord)] if values is not None else None)
                 if scores is None:
                     score = None
                 else:
                     # Step 2: Check if scores is an array or list (indexable)
                     if isinstance(scores, (np.ndarray, list)):
                         # Step 3: If scores is an array or list, index it
-                        score = int(scores[tuple(coord)])
+                        score = float(scores[tuple(coord)])
                     else:
                         # Step 4: If scores is a scalar, use it directly
-                        score = int(scores)
+                        score = float(scores)
 
                 results.append((coords_tuple, value, score))
         elif  isinstance(result, (int, np.integer)):
@@ -99,7 +99,7 @@ def check_variable_conditions(
                 coords_tuple = tuple(int(c) for c in combo + indices)
             else:
                 coords_tuple = indices
-            value = int(result)
+            value = float(result)
             results.append((coords_tuple, value))
     return results
